+++ src/aigora/curriculum_graph/application/assembling/assembler_errors.py
