+++ src/aigora/curriculum_graph/application/assembling/graph_assembler.py
