+++ src/aigora/curriculum_graph/application/graph_loader.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from typing import Any
 
 from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.application.assembling.graph_assembler import GraphAssembler
 
-from .graph_assembler import GraphAssembler
 from .graph_mapper import GraphMapper
 from .graph_parser import GraphParser
 from .graph_validator import GraphValidator
